Add Emotion component
