+++ components/Index.js
@@ -1,30 +1,6 @@
 import React from 'react';
-import styled, { css, injectGlobal } from 'react-emotion';
+import Emotion from './Emotion';
 
-injectGlobal`
-  body {
-    font-family: sans-serif;
-    margin: 0;
-  }
-`
+const Index = props => <Emotion />;
 
-const Container = styled('div')`
-  background: #ccc;
-  padding: 1rem;
-`
-
-const myStyle = css`
-  color: white;
-  margin: 0;
-`
-
-const Body = styled('p')({
-  color: 'red',
-})
-
-export default props => (
-  <Container>
-    <h1 className={myStyle}>Hello, World!</h1>
-    <Body>I am Body.</Body>
-  </Container>
-)
+export default Index;
